Add className input
by adding this param we can make this directive more versatile. Next
step woul be to have two classes: one for valid(wich is 'hide' now) and
another for invalid(wich is none now). Then we can apply this directive
to other elements, like a label or the input itself.